Arbeiten im Rahmen vom GL-Ertüchtigungsauftrag
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.iav.gllib@30427 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ src/de/bsvrz/iav/gllib/gllib/dav/GanglinieMQ.java
@@ -605,8 +605,8 @@ public class GanglinieMQ extends Ganglinie<Messwerte> {
 	 * @param bSplineOrdnung
 	 *            die neue Ordnung des B-Spline.
 	 */
-	public void setBSplineOrdnung(final int bSplineOrdnung) {
-		this.bSplineOrdnung = bSplineOrdnung;
+	public void setBSplineOrdnung(final long bSplineOrdnung) {
+		this.bSplineOrdnung = (int) bSplineOrdnung;
 	}
 
 	/**
